Add files via upload
Short description of code files:
I design the model in two parts: 
1.Selection process, which incorporates different competitive kernels (describing the ecology of competition) into the Lotka-Volterra dynamic.
2. Inheritance process: Explicit quantitative genetic models that describes how trait distributions are propagated as a result of reproduction. I employ two models- a) Quantitative genetic limit model (Assumes large no. of loci contribute to the phenotype, leading to robust Gaussian shaped genetic and trait variations). b) Shpak-Kondrashov hypergeometric distribution model (Assumes limited no. of biallelic loci additively contribute to the phenotype- Thus the offspring phenotype is sampled with Hypergeometric distribution from the parents gamete pool) which may not guarantee fixed trait variance and thus trait distribution shapes may deviate from normality under this model.

QG_test shows an example of a simulation under Quantitative genetic model of inheritance with a chosen set of parameters for comp. kernels and the relative strength of intraspecific competition. 
SK_test shows an example of a simulation under Shpak-Kondrashov model of inheritance with a chosen set of parameters for comp. kernels, no. of loci governing the trait and the relative strength of intraspecific competition. 

Dat_analysis codes are to be applied over the datasets generated with the simulation scheme over combinations of parameters of interest.